Set Kotlin module name to "insetter"
By default, the Kotlin compiler creates a kotlin_module file with the
name of the Gradle module. The resulting name is
"library_release.kotlin_module" for this library, which conflicts with
other libraries that also use the default generic "library" name. This
causes builds to fail when multiple such libraries are used in the same
project:

> A failure occurred while executing com.android.build.gradle.internal.tasks.Workers$ActionFacade
   > More than one file was found with OS independent path 'META-INF/library_release.kotlin_module'.

Set a unique Kotlin module name to fix the issue.